Disable cache in s-m-s, use unique "<anonymous>" "file name"
Disable cache when installing in source-map-support - we are keeping our
own cache, and so does ts-node, so while there is benefit from another
cache layer on s-m-s side, this causes problems with anonymous files, or
modules that are removed from require.cache and re-required.

Using "<anonymous iced>" instead of "<anonymous>" module.filename when
running code that does not have a file name allows iced3 and iced2 to
not clash if both are used within one node instance.